[BUGFIX] Render service exceptions as JSON
Adjusts the ``AbstractServiceController`` to catch all exceptions that
occur during request processing in order to convert them to JSON::

    {
      "error": {
        "code": 12345,
        "message": "Exception message",
        "details": "Optional message details",
        "referenceCode": "2015012315120876d8b2",
        "previous": {
          "code": 54321,
          "message": "Nested child exception"
        }
      }
    }

"details", "referenceCode" and "previous" are optional.

Note: This is a preliminary solution that improves error handling in
Backend. In the long term the exception handling of Flow will be
improved and this hotfix be deprecated.

Change-Id: Ifb5bf7c7537340b39fe4e84ea6c9b909648cffdd
Related: NEOS-108
Releases: master, 1.2