[TASK] Only store "lastVisitedNode" in session if it has been started
This change adjusts the ``NodeController`` to only store the
"lastVisitedNode" in the user session if it has been started before.
Usually that won't have any effect because an authenticated backend user
will start a session anyways. But it allows for replacing the default
``Typo3BackendProvider`` with a stateless implementation.

Change-Id: Id7c5bafd07f8a3483b74234cf73a18589303996a
Releases: master