[!!!][TASK] Reload content without reloading the whole page
The behavior of the node type property setting `reloadIfChanged`
to only reload the current element instead of reloading the whole
page. This makes the experience for the editors a lot smoother
especially on large pages or large elements, where the editor
would easily get confused by the position changing after reload.

This is breaking in case you rely on the whole page being reloaded
when a property of a single node is changed. To achieve the previous
behavior a new option called `reloadPageIfChanged` is introduced.

Change-Id: I247568a9a2a5515c64acf3cad6353fe4d3449d8f
Releases: master, 2.0
Resolves: NEOS-827