[!!!][TASK] ContentCollection overwrites node directly
Instead of creating another context variable the
ContentCollection TypoScript can directly overwrite ``node``
which reduces the amount of needed code and the context
dependencies of rendered content as it cannot use the
context variable. If a content element needs to know it's
ContentCollection it can use FlowQuery to find it.

This is breaking if you rely on the ``contentCollectionNode``
variable being set. You can retrieve the nearest
ContentCollection via FlowQuery.

Change-Id: I68f309ee84f4688446ae2c969b4cfaef84f52455
Releases: master