[FEATURE] NodeTreePrivilege
A privilege that prevents matching document nodes to appear in the
``Navigate Component``.
It also prevents editing of those nodes in case the editor navigates to
a node without using the Navigate Component (e.g. by entering the URL
directly).

Usage (in ``Policy.yaml``)::

 privilegeTargets:

   'TYPO3\Neos\Security\Authorization\Privilege\NodeTreePrivilege':

     'Some.Package:SomeIdentifier':
       matcher: 'isDescendantNodeOf("/sites/some-site/some-path")'

This would add a privilege that intercepts editing of the
``/sites/some-site/some-path`` node (and all of its child-nodes)
unless explicitly granted to the current role.

Change-Id: I87debf7b31d3b03dc5b0a4936d49e656fda20fc8
Resolves: NEOS-944
Releases: master