[BUGFIX] Fix rendering of HTML in translated string
This fixes the rendering of HTML in the backend template for
FlashMessages and in the History module.

Background:
Since the "Consistent escaping behavior" of Fluid (see
1e4cd0942dcf7b0726f3d3892bf8713cba89e9a4) the output of the translate
View Helper is escaped for security reasons, disallowing the usage of
HTML within translation labels unless the using side explicitly allows
it by using the ``{f:format.raw()}`` View Helper.

Change-Id: I9d1ca06c67aa1232a4f71481d902bc0deb2962c2
Releases: master