[!!!][BUGFIX] Account should not be available in the context
The current account entity should not be available in the context
but rather retrieved via global context or EEL helper.
Having it in the context leads to serialization of the account
which is not a good idea. Also depending on the cache configuration
you might get an old account.

This fix uses the new SecurityHelper introduced in accompanying
changes to get the account entity for rendering the backend information.
It drops account from the context.

This is marked breaking if you use the context variable ``${account}``
in your own TypoScript. You should instead use
``${Security.getAccount()}`` to retrieve it. Therefor you should also
remove all usage of account in safed contexts for uncached TypoScript
objects.

Change-Id: I22c88c13b4c57ef0ef6ae411cbacf783b8162cf6
Releases: master
Resolves: NEOS-1012