[!!!][TASK] Method to easily determine if backend rendering
To reduce exposure to implementation details two helper methods
have been added to find out if currently the backend is rendered
and if the live workspace is rendered.

This deprecates the TypoScript context variable ``editPreviewMode``
that was used inside of Neos to know what to render. The change is
breaking if you used this variable anywhere in your TypoScript.
You can seamlessly replace it by using
``${documentNode.context.currentRenderingMode.name}``.

Change-Id: I2c48f66716dfee756418f4fdb95b958a2bb24f01
Releases: master
Related: NEOS-845