[BUGFIX] uriPathSegment handling is slow and unreliable
This change fixes two issues with the mechanism which assures the
uniqueness of uriPathSegments among node siblings:

  1. Updating document nodes with lots of siblings was extremely slow
  2. Newly created nodes will get a non-human-friendly uriPathSegment if
     the proposed name was already taken by a sibling.

A severe side-effect was introduced with the change
``I2315d30256f7d291204363fdde7e761154cbc7e4``, making
adding/updating nodes extremely slow when a node has many siblings.

Multiplied by the check happening for every given node property change,
can lead to timeouts.

This changes the behavior to only do the necessary check if it's
actually needed, in case of new nodes, updating the ``uriPathSegment``
property or changing the path of an existing node.

Change-Id: I011a5196a8ab35c1b490c182ffdc955a1332f099
Releases: master, 1.2
Fixes: NEOS-1059