[BUGFIX] make error/exception handling more robust - allowing to reload in case of an error
Currently, if an error during rendering occurs, and then the user reloads
this document where the error occurs, the Neos backend breaks. This is because
some required context information is not added to neos-document-metadata,
breaking the page tree and the global workspace status indicator.

This is a regression introduced by  Ia8fa5228e772b0fee15921e02135c4899df39c0a

Change-Id: I39e6181868b234e53b29b37d0876826aac6efc9f
Releases: 1.2, master