[TASK] Make node migration compatible with json properties change
Because the image variant change and json properties change have to
be migrated in the right order we need to adjust the node migration
to not use the node model for getting the property. This is because
the node model returns NULL for the data as it is at this time
serialized where the model is assumed to contain json.

This change updates the migration so it can migrate correctly.

Change-Id: I5b6aa29a19cf2be411dadc757384f7b818413618
Releases: master
Resolves: NEOS-427