[BUGFIX] Form validation results and behavior
Form validation errors of backend submodules
were shown at the top of page with a wrong
offset and the problematic fields were not
highlighted.
Also several forms had a bad behavior when
pressing "enter" to submit and other small
bugs.

With this change the validation error list has
to be enabled by adding the
"validationResults" section and rendering
the "ValidationResults" viewhelper.
Most forms should not need this when all
errors are shown close to their fields.

Change-Id: Id2bd3194f81287a4c5e3853cbfd6668d2c47ac8e
Resolves: NEOS-264, NEOS-524, NEOS-1091
Releases: master