[BUGFIX] When a new site is created in Sites module, deactivate others
Disables existing Site Packages in case a new package is kickstarted
or a package is imported in the site management module.

This prevents a big variety of errors when people think the Demo Site
is inactive, but is still interacting with the their site because
Settings and NodeTypes are loaded.

Change-Id: I4f49447fae0ba5ac98acfcca83c32a97b5000be2
Resolves: NEOS-1108
Releases: master