[BUGFIX] Hide content element handles when page is selected
I085f9dd7a692e12155d2627b7da50911766ba108 Introduced a regression
that caused the content element handles to stay visible if
the page is selected while a content element was selected before.

This change prevents unneeded manupulation of the nodes array
in the node selection that caused this bug to happen.

Releases: master
Change-Id: Idf289bf0114b0703204aafd8550bafd65b579f37