[FEATURE] Introduce show action for Content Dimensions REST service
This change introduces a show action for the content dimensions service.
The new action allows for retrieval of presets of a single dimension or,
if a filter is specified, returns only those presets which are allowed
in combination with the given presets from other dimensions according
to possibly defined content dimension constraints.

Examples:

Retrieving only presets from the "language" content dimensions:

  http://localhost/neos/service/content-dimensions/language

Retrieving only those presets from the "language" dimensions which are
allowed in combination with the given "country" preset:

  http://localhost/neos/service/content-dimensions/language?chosenDimensionPresets[country]=de

Change-Id: I75c7d815e6957d6312c70cff8c159f0385aea0ad
Releases: master
Related: NEOS-1205