[!!!][TASK] Implement ContentCollection in pure TypoScript
The change `I06ba604f562685e3a1d5ee844a8adaec7eda20f5` was in fact
breaking if you need to overwrite properties of the actual
`Collection` renderer. Examples include the `itemName`,
`iterationName` or `itemRenderer`. Those must now be set inside the
`content` path of the `ContentCollection`.

So the following TypoScript::

  carouselItems = ContentCollection {
    nodePath = 'carouselItems'
    iterationName = 'carouselItemsIteration'
  }

Has to be changed to::

  carouselItems = ContentCollection {
    nodePath = 'carouselItems'
    content.iterationName = 'carouselItemsIteration'
  }

Change-Id: I4b26a5cf26a2ec4944f52a4442c61d7bcc8d51fc
Releases: master