[TASK] Return 404 on inaccessible nodes
The ``Frontend\\NodeController`` currently throws an exception when
trying to render a node that is protected by the ``ReadNodePrivilege``
because the ``$node`` argument can't be mapped.

This change adjusts this behavior by throwing an
``NodeNotFoundException`` (HTTP Status 404) if the requested node could
not be loaded.

Besides this change replaces some ``throwStatus()`` calls by proper
exceptions so that the regular error handling can kick in.

Background:

When the ``FrontendNodeRoutePartHandler`` resolves a request, the
``Security\\Context`` is not yet initialized (per design). So it can
resolve even protected nodes. But the ``PropertyMapper`` won't map
those leading to the exception.

Change-Id: Ied1577f94c980b06f2a0d27c49ac342349ebab69
Resolves: NEOS-1238
Releases: master, 2.0