[BUGFIX] ImageEditor should check if image was removed
In case the image gets removed from the editor an JavaScript
error was thrown due to not checking for a null value before
accessing properties on the editor value. On removal the editor
value property is set to null and then ``__updateValue()`` is
called. This now checks if value is a null value before using
any properties.

Change-Id: I013bea69c8d352d0aff043f92efe120087697b55
Releases: master, 2.0
Fixes: NEOS-1276