[BUGFIX] Give useful error messages in media management
When an error occurs in an action only a generic error
``FlashMessage`` is rendered. This won't help any users,
therefor we are adding some custom error handling first hiding
which action failed in ``Production`` context because no user
can use that information. But it adds any validation errors
as FlashMessages to the output to give users a hint what they
did wrong.

Change-Id: I4927f12fc3a5e111624f9385d251e4045e10c145
Releases: master, 2.0
Fixes: NEOS-264