[BUGFIX] Fix flushing of documents when parents subpage layout changes
The entry tag for documents is currently set to
`DescendantOf_{documentNode.identifier}` which will flush a document if
_any_ of the descendant nodes changes. This was originally  planned to
flush the document for parent layout change, but fails to do so and
causes lots of unnecessary cache flushes.

With the support of array cache entry tags we can do better by directly
referencing all parent document nodes in the tags.

Change-Id: Ic57980def129fa3af683b6468c8caa53ee42e7ea
Fixes: NEOS-1317
Releases: master, 2.0