[TASK] Show dialog on base workspace switch with pending content
This change introduces a dialog which is shown when the user tries to
select a different target workspace even though his personal workspace
contains changes.

To be on the safe side, we require the personal workspace to be empty
when rebasing it onto a different workspace.

Change-Id: I563b62b73738acd64c205055a690c7a7d75fed79
Releases: master
Related: NEOS-1239