[TASK] Make inital setup step clearer
Adds an explanation about the image libraries and slightly changes the
styling of the alerts to make clearer that it is OK to just have one of
the listed libraries. Also warns about the usage of GD in production due
to the memory issues experienced.

Additionally removes the "upload" test that in fact just tested if a
resource is accessible and if file_get_contents works on remote resources.
The second was a reported point of failure and due to the shutup "@"
operator nothing would be logged or displayed just the setup would say
your configuration is not ready to upload anything (which might or not
be true at this point).

We could reintroduce an automatic AJAX upload and display the upload
again to prove it worked.

Change-Id: Ie808e65c2a1db3c28c09f050656ae2ad62400aa3
Releases: master, 2.0