[!!!][FIX] create live workspace if neccessary during user workspace creation
Because the setup creates the user and his workspace before the site
is being imported, the user workspace doesn't get the live workspace
assigned as baseworkspace. This is not fixed during site import and so
the user can't login because his workspace doesn't contain any nodes.

Change-Id: I6584f1a4fd02e4a4046ae71b85727a0868501f64
Releases: master
Related: NEOS-1372