[TASK] Display node URI in link wizard, link editor & reference(s) editors
Display the node URI under the label in the Aloha link wizard, link editor and
in the reference(s) editors. The link wizard already shows the path, however
that is changed to the node URI since that's human readable as opposed to the path.

This commit also introduces the node.closestDocument view helper,
which returns the closest (itself included) node of type
``TYPO3.Neos:Document`` for a given node.

Change-Id: I8e4446e5ebcca51844f364c552c5edfa4e39bf8d
Resolves: NEOS-825
Releases: master, 2.0