[TASK] Don't trigger check for publishable nodes on vie create
This change improves the performance by omitting unnecessary events in
PublishableNodes.

The user interface would trigger updates of the publishable node count
as many times as there where vie entities on the page, iterating all
the entities as it was bound to the vie create event.

This is replaced by listening to internal events.

Change-Id: If459f199da3fa5253498311aae1613bcfef5d06e
Releases: master, 2.0