[BUGFIX] Update content correctly after applying inspector changes
When applying changes in the inspector for properties that require a
reload, the element wasn't updated correctly due to the change
``I18f86eeaa9f47737506fdb289833c2fe982bcb79`` changing the response
from a JSON formatted response to a redirect to a HTML response.

This change adjustes the Backbone sync to handle this change and
additionally makes the XHR request available for the success callback.

Change-Id: I50700e30034b33ac95d963f99f258aac1b832194
Related: NEOS-1368
Releases: master, 2.0